requirements.txt
asgiref==3.11.0
Django==6.0
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
django-meta==2.5.0
django-taggit==6.1.0
pillow==12.0.0
python-slugify==8.0.4
sqlparse==0.5.4
text-unidecode==1.3
tzdata==2025.2
